+++ src/backend/core/datatypes/Double2DayOfWeekFilter.h
@@ -43,7 +43,7 @@ public:
 protected:
 	//! Using typed ports: only double inputs are accepted.
 	bool inputAcceptable(int, const AbstractColumn *source) override {
-		return source->columnMode() == AbstractColumn::ColumnMode::Numeric;
+		return source->columnMode() == AbstractColumn::ColumnMode::Double;
 	}
 };
